+++ src/app/model/model.interfaces.ts
@@ -87,3 +87,11 @@ export interface IToken {
     exp: number;
     name: string;
 }
+
+export class Language {
+    constructor(
+        public code: string,
+        public name: string,
+        public resource: string
+    ) { }
+}
