--- src/app/model/utils/language.ts
@@ -1,7 +1,0 @@
-export class Language {
-    constructor(
-        public code: string,
-        public name: string,
-        public resource: string
-    ) { }
-}
